Improve appwrite auth (#15)
* feat: support client-side appwrite session

Expose Appwrite config client-side by prefixing env vars with VITE_ so they can be used in the client build.

Set Appwrite session cookie name to the standard Appwrite session cookie name so that if the cookie goes to Appwrite, it will be recognized.

* feat: update sign-out

Ensure the session is deleted from Appwrite.

* fix: set session cookie expiration to session expire

* fix: linting

---------

Co-authored-by: Steven Nguyen <stnguyen90@users.noreply.github.com>
Co-authored-by: Ariel Weinberger <weinberger.ariel@gmail.com>